requirements.txt
absl-py==1.4.0
altair==4.2.2
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
autokeras==1.1.0
autopep8==2.0.2
backcall==0.2.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
filelock==3.11.0
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
htbuilder==0.6.1
huggingface-hub==0.13.4
idna==3.4
imageio==2.27.0
importlib-metadata==6.1.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.10.0
keras-nlp==0.4.1
Keras-Preprocessing==1.1.2
keras-tuner==1.3.4
kiwisolver==1.4.4
kt-legacy==1.0.4
lazy_loader==0.2
libclang==16.0.0
lime==0.2.0.1
llvmlite==0.39.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
plotly==5.14.1
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.3
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy
semver==2.13.0
six==1.16.0
slicer==0.0.7
smart-open==6.3.0
smmap==5.0.0
st-annotated-text==4.0.0
stack-data==0.6.2
streamlit==1.17.0
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-text==2.10.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.3.21
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.4
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
